refactor(recursive): major refactoring, removes RecursiveWatcher interface
most liekely breaking some OSes, works on darwin and linux it seems (subject to some more extensive testing perhaps).

I figured that removing the RecursiveWatcher interface would be possible under the assumption: recursive watchers should be used iff an OS supports recursive watching of dirs. So basically, these two cases are disjoint: OS supports or does not support recursive watches ---> so we can split this by build tags. Of course if this assumption is false, then this change is completely off. But I believe it is sound.

- removes RecursiveWatcher interface
- splits tree creation to depend on OS
- adds additional args for Watch, Unwatch, Rewatch to support the potential of recursive watches (is_recursive, oldpath).
- update license file and adds headers in edited files
- minor changes to test suite limited to assertion and acting (such that it makes valid calls to the now changed interface).

---------

on-behalf-of: @SAP Simon Olander (simon.olander@sap.com)